Update dependencies and remove deprecated files
- Remove biome.json configuration from agentkit-core
- Add axios dependency to package.json
- Delete actions.ts file from src directory
- Update bun.lockb lockfile
