+++ agentkit-core/package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@0xgasless/smart-account": "latest",
+    "axios": "^1.7.9",
     "merkletreejs": "^0.4.1",
     "viem": "2",
     "zod": "^3.23.8"
